Trim migration-tools, add dependency in example
