+++ examples/utils/migration-tools/api-report/migration-tools.alpha.api.md
@@ -4,8 +4,6 @@
 
 ```ts
 
-export { EventEmitter }
-
 // (No @packageDocumentation comment for this package)
 
 ```
