+++ examples/utils/migration-tools/src/index.ts
@@ -3,22 +3,6 @@
  * Licensed under the MIT License.
  */
 
-// Provide EventEmitter from example-utils to avoid examples all directly depending on
-// a @fluid-internal package while EventEmitter support is finalized.
-import { EventEmitter } from "@fluid-internal/client-utils";
-export {
-	/**
-	 * @public
-	 */
-	EventEmitter,
-};
-
-export {
-	ContainerViewRuntimeFactory,
-	ViewCallback,
-	getDataStoreEntryPoint,
-	IFluidMountableViewEntryPoint,
-} from "./containerViewRuntimeFactory.js";
 export type {
 	DataTransformationCallback,
 	IAcceptedMigrationDetails,
@@ -28,22 +12,11 @@ export type {
 	IMigrationToolEvents,
 	IMigrator,
 	IMigratorEvents,
-	ISameContainerMigratableModel,
-	ISameContainerMigratableModelEvents,
-	ISameContainerMigrationTool,
-	ISameContainerMigrationToolEvents,
-	ISameContainerMigrator,
-	ISameContainerMigratorEvents,
 	IVersionedModel,
 	MigrationState,
-	SameContainerMigrationState,
 } from "./migrationInterfaces/index.js";
-export {
-	MigrationToolFactory,
-	SameContainerMigrationTool,
-	SameContainerMigrationToolInstantiationFactory,
-} from "./migrationTool/index.js";
-export { Migrator, SameContainerMigrator } from "./migrator/index.js";
+export { MigrationToolFactory } from "./migrationTool/index.js";
+export { Migrator } from "./migrator/index.js";
 export {
 	CreateModelCallback,
 	IAttachedMigratableModel,
@@ -62,20 +35,3 @@ export {
 	StaticCodeLoader,
 	TinyliciousModelLoader,
 } from "./modelLoader/index.js";
-export {
-	type IFluidMountableView,
-	type IProvideFluidMountableView,
-	MountableView,
-} from "./mountableView/index.js";
-export {
-	CollaborativeInput,
-	CollaborativeTextArea,
-	ICollaborativeInputProps,
-	ICollaborativeInputState,
-	ICollaborativeTextAreaProps,
-} from "./reactInputs/index.js";
-export {
-	ISharedStringHelperEvents,
-	ISharedStringHelperTextChangedEventArgs,
-	SharedStringHelper,
-} from "./SharedStringHelper.js";
