+++ examples/utils/migration-tools/src/migrator/index.ts
@@ -3,5 +3,4 @@
  * Licensed under the MIT License.
  */
 
-export { SameContainerMigrator } from "./sameContainerMigrator.js";
 export { Migrator } from "./migrator.js";
