--- examples/utils/migration-tools/src/mountableView/index.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export type { IFluidMountableView, IProvideFluidMountableView } from "./interface.js";
-export { MountableView } from "./mountableView.js";
