--- examples/utils/migration-tools/src/reactInputs/index.ts
@@ -1,14 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export {
-	ICollaborativeInputProps,
-	ICollaborativeInputState,
-	CollaborativeInput,
-} from "./CollaborativeInput.js";
-export {
-	CollaborativeTextArea,
-	ICollaborativeTextAreaProps,
-} from "./CollaborativeTextArea.js";
